% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\name{export2GeoLight}
\alias{export2GeoLight}
\title{Transforms twiligth data to match GeoLight requirements}
\usage{
export2GeoLight(twilights)
}
\arguments{
\item{twilights}{a dataframe generated by \code{\link{findTwilights}}.}
}
\value{
Return a data frame that matches GeoLight rquirements.
}
\description{
Transforms twGeos twilight data frames into format reguiered by 'GeoLight'
}

