% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\name{export2TAGS}
\alias{export2TAGS}
\title{Export data to match TAGS requirements}
\usage{
export2TAGS(tagdata, path = NULL, file = "exportTAGS", min = 0,
  max = 9999999999)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{path}{if full path is specified, the output will be writen as a txt file ready to import into TAGS}

\item{file}{name of txt file (without suffix)}

\item{min}{...}

\item{max}{...}
}
\value{
Return a data frame that matches rquirements for TAGS.
}
\description{
To pre- process geolocator datasets with TAGS it may be necessary 
to compress the data (i.e. remove repeated light levels).
}

