% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\name{thresholdCalibration}
\alias{thresholdCalibration}
\title{Search for twilight times}
\usage{
thresholdCalibration(
  twilight,
  rise,
  lon,
  lat,
  method = "log-normal",
  plot = TRUE
)
}
\arguments{
\item{twilight}{...}

\item{rise}{...}

\item{lon}{...}

\item{lat}{...}

\item{method}{...}

\item{plot}{...}
}
\value{
a vector with the _zero_ and the _median_ zenith angle as well as the parameters of the error distribution.
}
\description{
Calibration functon.
}
\details{
...
}
