% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\name{tsimageDeploymentLines}
\alias{tsimageDeploymentLines}
\title{tsimageDeployment}
\usage{
tsimageDeploymentLines(date, lon, lat, offset, zenith = 96, ...)
}
\arguments{
\item{date}{times as a vector of POSIXct.}

\item{lon}{longitude for deployment line.}

\item{lat}{latitude for deployment line.}

\item{offset}{the starting hour for the vertical axes.}

\item{zenith}{zenith angle for calculation of deployment line.}

\item{...}{additional arguments to pass to image.}
}
\description{
tsimageDeployment
}
