% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\docType{package}
\name{TwGeos-package}
\alias{TwGeos-package}
\title{Light-level geolocation data processing}
\description{
Basic data processing for data from light-level geolocation archival tags.
Provides facilities for importing and plotting data recorded by the
BAS tags, and in particular detecting times of twilight from the
recorded light data.
}
\author{
Simon Wotherspoon, Micheal Sumner and Simeon Lisovski.
}

