% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwGeos.R
\name{twilightAdjust}
\alias{twilightAdjust}
\title{Twilight Adjustment}
\usage{
twilightAdjust(twilights, interval, fixed = FALSE)
}
\arguments{
\item{twilights}{dataframe of twilight times.}

\item{interval}{the sampling interval.}

\item{fixed}{logical vector indicating twilights to be held fixed.}
}
\value{
a dataframe of bias correct twilight times.
}
\description{
Adjust twilight estimates for max light recording.
}
\details{
Some tags record the maximum light level observd in the preceding
sampling interval, and this behaviour biases naive estimates of
the times of sunset.  This function offsets sunset times by the
sampling interval to correct for this bias.
}
